Removed obsolete casting to "int".

+++ src/main/java/com/tightdb/lib/BinaryCursorColumn.java
@@ -10,17 +10,17 @@ public class BinaryCursorColumn<Cursor, Query> extends AbstractColumn<ByteBuffer
 
 	@Override
 	public ByteBuffer get() {
-		return cursor.rowset.getBinary(columnIndex, (int) cursor.getPosition());
+		return cursor.rowset.getBinary(columnIndex, cursor.getPosition());
 	}
 
 	@Override
 	public void set(ByteBuffer value) {
-		cursor.rowset.setBinary(columnIndex, (int) cursor.getPosition(), value);
+		cursor.rowset.setBinary(columnIndex, cursor.getPosition(), value);
 	}
 
 	@Override
 	public String getReadableValue() {
 		return "{binary}";
 	}
-	
+
 }
