+++ src/main/java/com/tightdb/lib/MixedCursorColumn.java
@@ -13,36 +13,36 @@ public class MixedCursorColumn<Cursor, Query> extends AbstractColumn<Mixed, Curs
 
 	@Override
 	public Mixed get() {
-		return cursor.rowset.getMixed(columnIndex, (int) cursor.getPosition());
+		return cursor.rowset.getMixed(columnIndex, cursor.getPosition());
 	}
-	
+
 	@Override
 	public void set(Mixed value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), value);
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), value);
 	}
 
 	public void set(String value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 	public void set(boolean value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 	public void set(long value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 	public void set(Date value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 	public void set(ByteBuffer value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 	public void set(byte[] value) {
-		cursor.rowset.setMixed(columnIndex, (int) cursor.getPosition(), TightDB.mixedValue(value));
+		cursor.rowset.setMixed(columnIndex, cursor.getPosition(), TightDB.mixedValue(value));
 	}
 
 }
